backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0.post1
httpx==0.25.2
pydantic>=2.7,<3
pydantic-settings>=2.1,<3
python-dotenv==1.0.1
